Replace source with reference 1.21.11 VulkanMod 0.5.10
Wholesale replacement of source tree, mixins, access widener,
assets, and fabric.mod.json from working 1.21.11 reference build.

+++ src/main/java/net/vulkanmod/render/engine/VkGpuBuffer.java
@@ -21,7 +21,7 @@ public class VkGpuBuffer extends GpuBuffer {
 
     Buffer buffer;
 
-    protected VkGpuBuffer(VkDebugLabel glDebugLabel, @Nullable Supplier<String> supplier, int usage, int size) {
+    protected VkGpuBuffer(VkDebugLabel glDebugLabel, @Nullable Supplier<String> supplier, int usage, long size) {
         super(usage, size);
         this.label = supplier;
